+++ packages/grafana-ui/src/visualizations/index.ts
@@ -1,0 +1,1 @@
+export { Graph } from './Graph/Graph';